requirements.txt
amqp==1.4.8
ansible==1.9.4
anyjson==0.3.3
appnope==0.1.0
arrow==0.7.0
attrs==16.0.0
backports-abc==0.4
bcrypt==3.1.0
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
billiard==3.3.0.22
BTrees==4.2.0
celery==3.1.19
certifi==2015.9.6.2
cffi==1.6.0
Chameleon==2.24
click==6.6
Coffin==2.0.1
configparser==3.5.0
cryptography==1.4
decorator==4.0.4
Django==1.9.4
django-celery==3.1.17
django-pagination==1.0.7
docx==0.2.4
ecdsa==0.13
entrypoints==0.2.2
enum34==1.1.6
f5-icontrol-rest==1.0.6
idna==2.1
ipaddress==1.0.17
IPy==0.83
ipykernel==4.5.0
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-core==4.2.0
kombu==3.0.32
lxml==3.6.0
MarkupSafe==0.23
martian==0.15
mechanize==0.2.5
mistune==0.7.3
MySQL-python==1.2.5
nbconvert==4.2.0
nbformat==4.1.0
notebook==4.2.3
numpy==1.9.2
packaging==15.3
Paginator==0.3.5
pam==0.1.4
pandas==0.17.1
paramiko==1.16.0
path.py==8.1.2
persistent==4.1.1
pexpect==4.0.1
pickleshare==0.5
Pillow==3.2.0
pinyin==0.3
portage==2.2.20.1
prettytable==0.7.2
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.5
pycparser==2.14
pycrypto==2.6.1
pygeoip==0.3.2
Pygments==2.1.3
pygobject==3.16.2
pygrok==0.7.1
pyPdf==1.13
pysphere==0.1.7
pysqlcipher==2.6.4
python-dateutil==2.4.2
python-docx==0.8.5
python-ldap==2.4.19
python-nmap==0.5.0
python-rrdtool==1.4.7
pytz==2015.7
pyvmomi==6.0.0.2016.6
pyxattr==0.5.3
pyxdg==0.25
PyYAML==3.11
pyzmq==15.4.0
redis==2.10.5
redis-py-cluster==1.1.0
regex==2016.4.8
requests==2.8.1
rrdtool==0.1.2
rsa==3.2.3
scons==2.3.5
shadowsocks==2.8.2
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.9.0
terminado==0.6
tornado==4.4.1
traitlets==4.0.0
transaction==1.4.4
uWSGI==2.0.13.1
waitress==0.8.10
WebOb==1.6.0
WebTest==2.0.21
xlrd==0.9.4
xlwt==1.0.0
z3c.autoinclude==0.3.6
